Check if the retrieved AES key is not null before proceeding
During the unlock of Encrypted Volume, the AES key returned from
computation was not checked in term of correctness. In case, when the
AES key's DO was removed from the smart card (e.g. due to its factory
reset via CCID interface), the error code was not checked and empty,
zeroed buffer was used as an AES key instead.

Presented patch does not allow to unlock and use Encrypted Volume in
case of missing AES DO from smart card.